enigma2: Use servicemp3 from openatv-dev branch.
This is a big change for servicemp3.
It changes servicemp3.cpp and servicemp3.h of openpli master branch to the
openatv dev branch.
To adapt enigma2 to the latest gstreamer.

It includes this bunch of commits.
(most of them are from june 1st 2016 till february 24th 2017)

https://github.com/openatv/enigma2/commits/DEV/lib/service/servicemp3.cpp

except these two commit(as far as possible).

https://github.com/OpenPLi/enigma2/commit/b4841d0609e1f6f459d2be1fb0a9171873de06c8

https://github.com/OpenPLi/enigma2/commit/8bd6209812275e7b439e00884148213425afcdce

It's an indiscriminate manner somewhat, but inherent when you're between two worlds.
I have decided it's better than forking the enigma2 repo of openpli and cherry picking a bunch of commits.

It solves the pause/unpause issue e.g. when playing physicall cd.
With original servicemp3, the player is totally lost in position when pause for the second time.
With this patch pause/unpause works flawless and probably for more media.

But pause a wav or mp3 file from hdd, the file still stops when pause.

It's still a test, but it has improvements.

Thanks to christophecvr and all other openatv dev's.