layer.conf: silence warnings LAYERSERIES_COMPAT
http://cgit.openembedded.org/bitbake/commit/?id=cca81e33b58c390dcf5cc3a31555a43b79177166

LAYERSERIES_COMPAT_openembedded-layer in its conf/layer.conf file to list the core layer names it is compatible with.
LAYERSERIES_COMPAT_filesystems-layer in its conf/layer.conf file to list the core layer names it is compatible with.
LAYERSERIES_COMPAT_multimedia-layer in its conf/layer.conf file to list the core layer names it is compatible with.
LAYERSERIES_COMPAT_networking-layer in its conf/layer.conf file to list the core layer names it is compatible with.
LAYERSERIES_COMPAT_meta-python in its conf/layer.conf file to list the core layer names it is compatible with.