Run package-index after do_rootfs
Not at the very beginning of the build.
This prevent to require to run the build twice for image packages before updating the box.

For feed packages we still need to run the build twice in some situations.
It depends if package-index is executed before or after the build of the feed is done.
In case of before, we need to build the feed twice (in case of e.g. building from scratch) before updating the box.

Simply changing  do_package_index[depends] with a bbappend didn't work.
So copy the recipe from openembedded-core to this layer, add openpli-enigma2-image:do_rootfs to depends and point to it.