enigma2: add 'update-about-screen' patch
Get rid of flash type, version and revision, since we don't have versioninfo.bb
and we don't use a revision number.
Also we known it's a normal image and not a small flash image.
Use the old way to display version number (after image name).
Add 'unofficial'  to image name to make it clear it's not an official Open Vision image.